requirements.txt
line-bot-sdk

python-aiml

Flask
future
gunicorn
itsdangerous
Jinja2
MarkupSafe
psycopg2
requests
urllib3
Werkzeug
beautifulsoup4

google
googlemaps
googletrans
pyowm
feedparser
PyMySQL